+++ src/compas/files/stl.py
@@ -14,13 +14,6 @@ __all__ = [
 
 
 class STL(object):
-
-    def __init__(self, filepath, precision=None):
-        self.reader = STLReader(filepath)
-        self.parser = STLParser(self.reader, precision=precision)
-
-
-class STLReader(object):
     """Standard triangle library format.
 
     See Also
@@ -28,6 +21,13 @@ class STLReader(object):
     * http://paulbourke.net/dataformats/stl/
 
     """
+
+    def __init__(self, filepath, precision=None):
+        self.reader = STLReader(filepath)
+        self.parser = STLParser(self.reader, precision=precision)
+
+
+class STLReader(object):
 
     def __init__(self, filepath):
         self.filepath = filepath
@@ -52,8 +52,7 @@ class STLReader(object):
                 self.read_binary()
         except Exception:
             # raise if it was already detected as binary, but failed anyway
-            if is_binary:
-                raise
+            if is_binary: raise
 
             # else, ascii parsing failed, try binary
             is_binary = True
@@ -125,7 +124,7 @@ class STLReader(object):
 
             # no known line start matches, maybe not ascii
             elif not parts[0].isalnum():
-                raise RuntimeError('File is not ASCII')
+                raise Exception('File is not ASCII')
 
         return facets
 
@@ -197,9 +196,9 @@ class STLParser(object):
 
     def __init__(self, reader, precision=None):
         self.precision = precision
-        self.reader = reader
-        self.vertices = None
-        self.faces = None
+        self.reader    = reader
+        self.vertices  = None
+        self.faces     = None
         self.parse()
 
     def parse(self):
@@ -237,6 +236,7 @@ if __name__ == "__main__":
     from compas_viewers import MeshViewer
     from compas.utilities import download_file_from_remote
     from compas.topology import connected_components
+
 
     source = 'https://raw.githubusercontent.com/ros-industrial/abb/kinetic-devel/abb_irb6600_support/meshes/irb6640/visual/link_1.stl'
     filepath = os.path.join(compas.APPDATA, 'data', 'meshes', 'ros', 'link_1.stl')