+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- File reading functions for ascii files in `compas.files` has moved from the individual reader classes to a new parent class, `BaseReader`.
 - Added inherited methods to class docs.
 - Added data structure mixins to the docs.
 - Added `data` and `from_data` to `compas.geometry.Polyhedron`